Reorganized features folder structure
