+++ app/features/components/List.tsx
@@ -3,7 +3,7 @@
 import { FC } from 'react';
 import FeatureCard from '../components/Card';
 
-const List: FC<{ features: any[] }> = ({ features }) => {
+const List: FC<{ features: { title: string; icon: string; description: string;}[] }> = ({ features }) => {
     return (
         <div className="container mx-auto w-full p-5 mt-8 mb-6">
             <div className="grid grid-cols-1 md:grid-cols-3 mb-11 mt-8 gap-8 sm:gap-10">
@@ -14,7 +14,6 @@ const List: FC<{ features: any[] }> = ({ features }) => {
                             title={feature.title}
                             icon={feature.icon}
                             description={feature.description}
-                            slug={feature.slug}
                             width={"w-11/12"}
                         />
                     </div>
@@ -23,6 +22,5 @@ const List: FC<{ features: any[] }> = ({ features }) => {
         </div>
     );
 }
-
 
 export default List;
