+++ app/features/omni-channel/page.tsx
@@ -1,0 +1,44 @@
+import type { Metadata } from "next";
+import OmniChannelFeature from "./omni-channel";
+
+
+export const metadata: Metadata = {
+  title: "Proof of X | OmniChannel",
+  description:
+    "",
+  openGraph: {
+    title: "Proof of X | OmniChannel",
+    description:
+      "",
+
+    url: "https://pox.me/dashboard/omni-channel",
+    siteName: "Proof of X",
+    images: [
+      {
+        url: "/images/home-hero.jpg",
+        width: 1800,
+        height: 1600,
+        alt: "A cool fingerprint image",
+      },
+    ],
+    locale: "en_US",
+    type: "website",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "Proof of X | OmniChannel",
+    description:
+      "",
+
+    creator: "@proofxme",
+    images: ["/images/home-hero.jpg"],
+  },
+};
+
+export default function OmniChannelFeaturePage() {
+  return (
+    <>
+      <OmniChannelFeature />
+    </>
+  );
+}
